real-time-chat-moderator/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
joblib==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
